Moved duplicate code (DuplexGetter) into Pre-compiled
